EES-3777 Split meta view model builders into separate files in Utils namespace
